Fixed the /chat endpoints

+++ parley-service-jpa/src/main/java/org/sidequest/parley/service/UserService.java
@@ -34,7 +34,7 @@ public class UserService {
     }
 
     public User getUser(Long userId) {
-        return userRepository.findById((long) userId)
+        return userRepository.findById(userId)
                 .map(UserMapper.INSTANCE::toModel)
                 .orElse(null);
     }
